refactor: split the two files over rust-file-max-lines (Dog2)
Both files exceeded the self-lint's 2000-line rust-file-max-lines
threshold (warning-level), so the dogfood wasn't fully green. Pure code
movement, no logic change:

- alint-dsl/src/lib.rs (2335, mostly a ~1500-line test module): the
  #[cfg(test)] mod tests moved to src/tests.rs (mod tests;). lib.rs is
  now 836 lines. The tests use `use super::*`, semantically identical as
  a file module.

- xtask/src/docs_export.rs (2131, mostly code): the ~213-line test
  module moved to docs_export/tests.rs, and the self-contained ~180-line
  drift-gate count parsers to docs_export/counts.rs. The 5
  count_canonical_* entry points are pub(super); write_manifest calls
  them qualified (counts::...). strip_nested_braces is pub(super) for the
  one CLI-reference test that shares it. docs_export.rs is now 1738.

Verified: alint-dsl (120) + xtask (85) tests pass, fmt + clippy clean,
the dogfood is now fully green (All 41 rules passed, no warnings), and
docs-export still emits the correct manifest counts (89/22/11/8/12).
